Version3/requirements.txt
streamlit
google-genai
python-dotenv
pillow
watchdog

# API 服务器依赖
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
